Fixed inverse transform/improved naming convention
The inverse transformation or 'undo' button did not work correctly and
is fixed by using the inverted 2x2 matrix.

Naming convention made it hard to read the code. Added clearer ui element
names.

Affine transformation was confusing, normally x' = A x + b, where A = [a
b; c d] and b = [tx; ty]. b and c are not the 'rotation' parameters.